ENH: Relocate orphaned domain code to Liver/_LegacyLiverLogic.py
Move the LiverLogic class (signed-Maurer distance maps, Bezier surface
fitting, elliptic Fourier descriptors, polydata utilities) out of
Liver/Liver.py into a private sibling module. The leading underscore
signals "awaiting full relocation to the per-stage modules" tracked
in issue #437.

Liver.py re-exports LiverLogic so existing call sites continue to
work; commit 3 of this stack (LiverWidget.setup() rewrite) drops the
re-export once the shell stops instantiating it.

The bezier-characterisation pytest fixture is rewired to load
_LegacyLiverLogic.py directly so the test no longer depends on the
soon-to-shrink Liver.py shell.

Per ADR-0023 §"Shell composition (Option H)": the Liver shell holds
no domain logic — only composition + navigation. Partial step in
that direction; full relocation continues in #437.
